+++ mcp_tools/cloud_audit/prowler.py
@@ -1,0 +1,38 @@
+# mcp_tools/cloud_audit/prowler.py
+
+from typing import Dict, Any
+
+def register_prowler_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def prowler_scan(provider: str = "aws", profile: str = "default", region: str = "", checks: str = "", output_dir: str = "/tmp/prowler_output", output_format: str = "json", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Prowler for comprehensive cloud security assessment.
+
+        Args:
+            provider: Cloud provider (aws, azure, gcp)
+            profile: AWS profile to use
+            region: Specific region to scan
+            checks: Specific checks to run
+            output_dir: Directory to save results
+            output_format: Output format (json, csv, html)
+            additional_args: Additional Prowler arguments
+
+        Returns:
+            Cloud security assessment results
+        """
+        data = {
+            "provider": provider,
+            "profile": profile,
+            "region": region,
+            "checks": checks,
+            "output_dir": output_dir,
+            "output_format": output_format,
+            "additional_args": additional_args
+        }
+        logger.info(f"☁️  Starting Prowler {provider} security assessment")
+        result = hexstrike_client.safe_post("api/tools/prowler", data)
+        if result.get("success"):
+            logger.info(f"✅ Prowler assessment completed")
+        else:
+            logger.error(f"❌ Prowler assessment failed")
+        return result
