+++ mcp_tools/cloud_container/cloud_and_container_security.py
@@ -5,72 +5,6 @@ from typing import Dict, Any
 def register_cloud_and_container_security_tools(mcp, hexstrike_client, logger):
 
     @mcp.tool()
-    def prowler_scan(provider: str = "aws", profile: str = "default", region: str = "", checks: str = "", output_dir: str = "/tmp/prowler_output", output_format: str = "json", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Prowler for comprehensive cloud security assessment.
-
-        Args:
-            provider: Cloud provider (aws, azure, gcp)
-            profile: AWS profile to use
-            region: Specific region to scan
-            checks: Specific checks to run
-            output_dir: Directory to save results
-            output_format: Output format (json, csv, html)
-            additional_args: Additional Prowler arguments
-
-        Returns:
-            Cloud security assessment results
-        """
-        data = {
-            "provider": provider,
-            "profile": profile,
-            "region": region,
-            "checks": checks,
-            "output_dir": output_dir,
-            "output_format": output_format,
-            "additional_args": additional_args
-        }
-        logger.info(f"☁️  Starting Prowler {provider} security assessment")
-        result = hexstrike_client.safe_post("api/tools/prowler", data)
-        if result.get("success"):
-            logger.info(f"✅ Prowler assessment completed")
-        else:
-            logger.error(f"❌ Prowler assessment failed")
-        return result
-
-    @mcp.tool()
-    def trivy_scan(scan_type: str = "image", target: str = "", output_format: str = "json", severity: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Trivy for container and filesystem vulnerability scanning.
-
-        Args:
-            scan_type: Type of scan (image, fs, repo, config)
-            target: Target to scan (image name, directory, repository)
-            output_format: Output format (json, table, sarif)
-            severity: Severity filter (UNKNOWN,LOW,MEDIUM,HIGH,CRITICAL)
-            output_file: File to save results
-            additional_args: Additional Trivy arguments
-
-        Returns:
-            Vulnerability scan results
-        """
-        data = {
-            "scan_type": scan_type,
-            "target": target,
-            "output_format": output_format,
-            "severity": severity,
-            "output_file": output_file,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting Trivy {scan_type} scan: {target}")
-        result = hexstrike_client.safe_post("api/tools/trivy", data)
-        if result.get("success"):
-            logger.info(f"✅ Trivy scan completed for {target}")
-        else:
-            logger.error(f"❌ Trivy scan failed for {target}")
-        return result
-    
-    @mcp.tool()
     def scout_suite_assessment(provider: str = "aws", profile: str = "default",
                               report_dir: str = "/tmp/scout-suite", services: str = "",
                               exceptions: str = "", additional_args: str = "") -> Dict[str, Any]:
