+++ mcp_tools/container_scan/trivy.py
@@ -1,0 +1,37 @@
+# mcp_tools/container_scan/trivy.py
+
+from typing import Dict, Any
+
+def register_trivy_tool(mcp, hexstrike_client, logger):
+
+    @mcp.tool()
+    def trivy_scan(scan_type: str = "image", target: str = "", output_format: str = "json", severity: str = "", output_file: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Trivy for container and filesystem vulnerability scanning.
+
+        Args:
+            scan_type: Type of scan (image, fs, repo, config)
+            target: Target to scan (image name, directory, repository)
+            output_format: Output format (json, table, sarif)
+            severity: Severity filter (UNKNOWN,LOW,MEDIUM,HIGH,CRITICAL)
+            output_file: File to save results
+            additional_args: Additional Trivy arguments
+
+        Returns:
+            Vulnerability scan results
+        """
+        data = {
+            "scan_type": scan_type,
+            "target": target,
+            "output_format": output_format,
+            "severity": severity,
+            "output_file": output_file,
+            "additional_args": additional_args
+        }
+        logger.info(f"🔍 Starting Trivy {scan_type} scan: {target}")
+        result = hexstrike_client.safe_post("api/tools/trivy", data)
+        if result.get("success"):
+            logger.info(f"✅ Trivy scan completed for {target}")
+        else:
+            logger.error(f"❌ Trivy scan failed for {target}")
+        return result
